% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSankey.R
\name{makeSankey}
\alias{makeSankey}
\title{Make a Sankey Plot}
\usage{
makeSankey(d, white.horizontal = F, white.vertical = F, title = NA)
}
\arguments{
\item{d}{data fram}

\item{white.horizontal}{defaults to F}

\item{white.vertical}{defaults to F}

\item{title}{defaults to NA which uses d[1,1] as title}
}
\description{
This functions takes in a data frame and makes a Sankey Plot
It assumes that the first column contains risk factors
and following columns are each numerical values for each factor for that year
}
\details{
This function can handle any number of years and/or risk factors
The presense of horizontal and vertical gaps can be adjusted
}
